requirements.txt
torch
torchvision
torchaudio
opencv-python
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
flask
tqdm
transformers
librosa
soundfile
